requirements.txt
Flask==2.1.2
pandas==1.4.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.56.0
serpapi==2.0.0
openai==0.27.0
